requirements.txt
ldap3==2.4
pyasn1==0.4.2
PyYAML==3.12
